+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+ - Fixed an issue in IcdUriBuilder where relative pathes were not being built into a valid URI.
 
 ## [15.2.0] - 2021-08-18
 ### Added